restructure project with consitent naming
